BREAKING: Queries struct stop using interface for cache.
DCache upgraded to return *DCache instead of the Cache
interface in its New() function at some point. Since then,
User might pass a pointer of (nil)(*dcache.DCache) to Queries,
which is not a nil interface but an interface holding an nil
poitner. This makes the internal nil check failed to catch.

There are many solutions:
1. Keep it like this, and user must be careful not to pass nil pointer.
   Because go is not null-safe, there is no compilation-time check for
   this. Then if unit-test coverage is low, user will run into this issue
   on production, and panic.
2. DCache returns interface: no not supported. Besides it may still have
   the above issue, just unlikely.
3. Abandon cache interface, use concrete struct. The tradeoff is that
   we can no longer inject other cache implementation, whichi might be
   a useful feature during testings. However, I have never seen this
   in real life so it might be YAGNI.

So we choosed (3) as the final solution.